Create configures for ANUSPLIN

Format the input data and generate the configuration file required for ANUSPLIN interpolation.

anusplin_params(
  dat,
  basename,
  range,
  file.alt,
  unit = 0,
  res = 0.25,
  width = 7,
  alt = "cov",
  lim.lon = "auto",
  lim.lat = "auto",
  lim.alt = "auto",
  cvt.lon = c(0, 1),
  cvt.lat = c(0, 1),
  cvt.alt = c(1, 1),
  cvt.coef = 1000,
  trans.dep = 0,
  order = 3,
  err.wgt = 0,
  optimize = 1,
  smooth = 1,
  type.mask = 0,
  file.mask = NULL,
  type.alt = 2,
  type.grd = 2,
  missing = -9999,
  err.cov = 2,
  grid.pos = 1,
  essential = T
)

Arguments

dat

A data.frame or data.table to interpolate, for the colnames, lon, lat must be included, site and alt are optional, the others are variable names.

basename

Basename for all output files, extensions should not be included.

range

Range of interpolation grid (xmin, xmax, ymin, ymax).

file.alt

When type.alt is not 0, this is the input grid file name; otherwise a manual constant.

unit

The unit of dat, a non-negative integer, possible values are:

0

undefined (default)

1

meteres

2

feet

3

kilometers

4

miles

5

degrees

6

radians

7

millimetres

8

megajoules

res

The grid resolution (degree).

width

The fixed width of numbers in formatted data.

alt

Type of elevation was treated, possible values are:

NULL

no use of elevation

cov

considered as independent covariates (default)

spl

considered as independent spline variables

lim.lon

A vector containing lower and upper limits, auto (default) meant to use the minimum and maximum values in the data, or set manually. Data points outside these limits, augmented by margins, are ignored.

lim.lat

Same as lim.lon, but for longitude.

lim.alt

Same as lim.lon, but for altitude.

cvt.lon

Transformation and scale factor (default is 1) of longitude.
Real Value = (Table Value) * (Scale Factor)
The possible transformations are:

0

no transformation (default)

1

x/a

2

ax

3

a·log(x+b)

4

(x/b)^a

5

a·exp(x/b)

6

a·tanh(x/b)

7

anisotropy angle in degrees

8

anisotropy factor - in the direction specified by the anisotropy angle

cvt.lat

Same as cvt.lon, but for latitude.

cvt.alt

Same as cvt.lon, but for altitude.

cvt.coef

Parameters used for transformation, one or two real numbers.

trans.dep

Dependent variable transformation, possible values are:

0

no transformation (default)

1

fit surface to natural logarithm of the data values

2

fit surface to the square root of the data values

5

occurrence – transform data values by setting all positive value to 1.0 and ignoring all negative values

order

Order of spline, a positive integer.

err.wgt

Number of relative error variances, a non-negative integer, possible values are:

0

data points uniformly weighted for each surface (default)

1

the same weighting is applied to each surface

Number of surfaces

a different weighting is applied to each surface

optimize

Optimization directive, a non-negative integer, possible values are:

0

common smoothing parameter for all surfaces

1

common smoothing directive for all surfaces (default)

2

different smoothing directive for each surface

smooth

Smoothing directive for each surface, a non-negative integer, possible values are:

0

fixed smoothing parameter - supply value

1

minimise GCV (default)

2

minimise true mean square error using supplied error standard deviation estimate

3

fixed signal - supply value

4

minimise GML

type.mask

Mode of mask grid, a non-negative integer, possible values are:

0

mask grid not supplied (default)

1

generic mask grid

2

Arc/Info mask grid

3

Idrisi mask grid

file.mask

Filename of mask grid, only valid if type.mask set to positive integer.

type.alt

Mode of the independent variable, possible values are:

0

user supplied constant

1

user supplied grid in generic row format with the same size as the grid being calculated

2

user supplied Arc/Info grid with same size as the grid being calculated (default)

3

user supplied Idrisi image with the same size as the grid being calculated

type.grd

Same as type.mask, but for interpolated grid.

missing

Filling of missing values.

err.cov

test

grid.pos

Grid position option, a non-negative integer, possible values are:

0

grid points at cell corners (default)

1

grid points at cell centres

essential

If True, only export essential process files, large residual file, optimisation parameters file, data list file and validation data file are ignored.

Value

a list with three components:

data

formatted data.table of dat

splina

a vector containing splina parameters

lapgrd

a vector containing lapgrd parameters

Examples

data(TempBrazil)
colnames(TempBrazil) <- c("lon", "lat", "temp")
anusplin_params(TempBrazil, "TempBrazil", c(70, 140, 15, 55), "dem.txt", alt = NULL)
#> Error in anusplin_params(TempBrazil, "TempBrazil", c(70, 140, 15, 55),     "dem.txt", alt = NULL): could not find function "anusplin_params"