Weights for spatial interpolation

Weights for spatial interpolation

cal_weight(
  points,
  range,
  res = 0.5,
  cdd = 450,
  m = 4,
  nstation.max = 8,
  nstation.min = 3,
  wFUN = c("wFUN_adw", "wFUN_idw"),
  .progress = "none",
  ...
)

cal_weight_sf(
  points,
  range = NULL,
  res = 0.25,
  cdd = 450,
  m = 4,
  nstation.max = 8,
  nstation.min = 3,
  wFUN = c("wFUN_adw", "wFUN_idw"),
  .progress = "none",
  ...
)

Arguments

points

A matrix (N,2) with longitude and latitude of points of data observed

range

`[xmin, xmax, ymin, ymax]`

res

the grid resolution (degree)

cdd

the correlation decay distance (km) for `adw` and searching radius for others, default (450).

m

distance weight (default 4)

nstation.max

Number of maximum stations used per point for interpolation, (default 8).

nstation.min

Number of minimum stations used per point for interpolation, (default 3).

.progress

name of the progress bar to use, see create_progress_bar

...

other parameters to [plyr::ldply]

Value

A data.frame with longitude, latitude and interpoled points